src/auth/requirements.txt
Flask==2.2.3
Flask_MySQLdb==1.0.1
PyJWT==2.6.0
